+++ p4-fusion/CMakeLists.txt
@@ -11,7 +11,9 @@ find_package(OpenSSL)
 add_executable(p4-fusion ${Headers} ${Sources}
         commands/labels_result.cc
         commands/labels_result.h
-        commands/label_result.h)
+        commands/label_result.h
+        labels_conversion.cpp
+        labels_conversion.h)
 
 target_include_directories(p4-fusion PUBLIC
         ../${HELIX_API}/include/
